requirements.txt
pdfminer.six
python-magic